Finished bhud for the moment
So, here is the new BHUD!
Maybe there will be more features later, but for now this update is cool
enough! :D

- Completely changed the design (Made some texts more readable)
- Huge code-improvements for better performance and a better overview
- Improved the cl_drawhud client-command
- Added SQL-Settings to change the bHUD-Settings (type in chat
"!bhud_help" or something else, which begins with "!bhud_" to see all
available settings!
- Added custom images which will be downloaded if someone joins the
server, where bhud is installed (it is not required to have a client
copy to see all images!)
- Each image, there are 6 of them, has a size of about 500 bytes, so
that shouldn't be a problem
- Fixed the problem, that bhud didn't work in singleplayer
- Highly improved the time-system. This works now faster than before and
looks a bit smarter!